move visualisations into own folders
