+++ frontend/src/app/visualization/ngram/ngram.component.html
