+++ frontend/src/app/visualization/ngram/ngram.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { commonTestBed } from '../common-test-bed';
+import { commonTestBed } from '../../common-test-bed';
 import { NgramComponent } from './ngram.component';
 
 describe('NgramComponent', () => {
