+++ frontend/src/app/visualization/wordcloud/wordcloud.component.scss
