+++ frontend/src/app/visualization/wordcloud/wordcloud.component.ts
@@ -3,8 +3,8 @@ import { Component, ElementRef, EventEmitter, Input, OnChanges, OnDestroy, OnIni
 import * as cloud from 'd3-cloud';
 import * as d3 from 'd3';
 
-import { AggregateResult, visualizationField, QueryModel, Corpus, freqTableHeaders } from '../models/index';
-import { DialogService, SearchService, ApiService } from '../services/index';
+import { AggregateResult, visualizationField, QueryModel, Corpus, freqTableHeaders } from '../../models/index';
+import { DialogService, SearchService, ApiService } from '../../services/index';
 import { BehaviorSubject, Observable } from 'rxjs';
 
 @Component({
